requirements.txt
falcon
uwsgi
pyyaml
